requirements.txt
# application
kafka-python
websockets

# testing
black
flake8
isort
mypy
pytest
